Fix >256 terrain gen block ID in Mo' Creatures wyvern biome
These IDs must be below 256 for proper terrain generation,
otherwise they are truncated to 8 bits.. which was causing saplings
to spawn on the surface, causing a StackOverflowError in block
updates, causing an out of memory error, causing the crashreporter
to load, but since its out of memory causing the error:
java.lang.NoClassDefFoundError: net/minecraft/world/CallableLvl1

https://f.cloud.github.com/assets/2968210/350028/6e8be65c-a002-11e2-9903-986250055020.png
Not an MCPC+ bug but filed https://github.com/MinecraftPortCentral/MCPC-Plus/issues/707

ModAnalyzer picked a high ID, but as the comments say, <256 is
required. However there is no ID restriction annotation, as used
for example in ExtraBiomesXL for ID Resolver:
https://github.com/ExtrabiomesXL/ExtrabiomesXL/blob/master/common/extrabiomes/utility/IDRestrictionAnnotation.java
so we have to fix this ID manually.

Now you can properly generate wyvern biomes.